requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0
termcolor>=1.1.0

# AI/ML dependencies
openai>=1.0.0  # For DeepSeek API (OpenAI compatible)

# Technical analysis
pandas-ta>=0.3.14b

# Crypto/Trading
pybit>=5.0.0  # Official Bybit Python SDK (alternative to our custom client)

# Solana dependencies (for existing functionality)
solders>=0.18.0
solana>=0.30.0

# Optional: For enhanced functionality
schedule>=1.2.0  # For advanced scheduling
colorama>=0.4.4  # Enhanced terminal colors
matplotlib>=3.5.0  # For plotting (optional)
seaborn>=0.11.0  # For enhanced plotting (optional)